+++ CelestiaWorks/source/vulkanAPI/Pipeline.cpp
@@ -114,8 +114,8 @@ VkViewport celestia::Pipeline::createViewport()
 	VkViewport viewport{};
 	viewport.x = 0.0f;
 	viewport.y = 0.0f;
-	viewport.width = swapChain.extent.width;
-	viewport.height = swapChain.extent.height;
+	viewport.width = static_cast<float>(swapChain.extent.width);
+	viewport.height = static_cast<float>(swapChain.extent.height);
 	viewport.minDepth = 0.0f;
 	viewport.maxDepth = 1.0f;
 	return viewport;
